:building_construction: Change Components Structure
