+++ src/components/Header/Header.less
@@ -1,0 +1,15 @@
+@import '../../style/themes/default';
+
+.header {
+  background: @base-white;
+  padding: 1.6rem;
+  color: @text-color;
+  border-bottom: 1px solid @white-3;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+}
+
+.tool-container {
+  display: flex;
+}
