+++ src/components/SiderMenu/SiderMenu.js
@@ -3,7 +3,7 @@ import { Layout, Menu, Icon } from 'antd';
 import pathToRegexp from 'path-to-regexp';
 import Link from 'umi/link';
 import Debounce from 'lodash-decorators/debounce';
-import styles from './index.less';
+import styles from './styles.less';
 
 const { Sider } = Layout;
