+++ src/components/TagInput/index.js
@@ -1,0 +1,3 @@
+import TagInput from './TagInput';
+
+export default TagInput;
