+++ src/components/TagInput/styles.less
@@ -1,14 +1,6 @@
-@import "../../style/themes/default";
+@import '../../style/themes/default';
 
 :global {
-  .ant-tag-checkable {
-    background: @base-white;
-    border: 1px solid @transBlack-6;
-  }
-  .ant-tag-checkable-checked {
-    background-color: @primary-color;
-    border: 1px solid @primary-color;
-  }
   .ant-tag {
     height: 24px;
     display: flex;
@@ -16,6 +8,15 @@
     align-items: center;
     padding: 0 12px;
     margin-right: @padding-md;
+
+    &-checkable {
+      background: @base-white;
+      border: 1px solid @transBlack-6;
+    }
+    &-checkable-checked {
+      background-color: @primary-color;
+      border: 1px solid @primary-color;
+    }
   }
   ::-webkit-scrollbar {
     width: 12px;
