+++ src/components/TreeList/index.js
@@ -1,0 +1,4 @@
+import _SearchTree from './SearchTree';
+
+
+export const SearchTree = _SearchTree;
