+++ src/components/index.js
@@ -1,0 +1,12 @@
+import _TreeList from './TreeList';
+// import _SiderMenu from './SiderMenu';
+import _EditableItem from './EditableItem';
+import _Login from './Login';
+import _ListInput from './ListInput';
+import _TagInput from './TagInput';
+
+export const TreeList = _TreeList;
+export const EditableItem = _EditableItem;
+export const Login = _Login;
+export const ListInput = _ListInput;
+export const TagInput = _TagInput;
