+++ src/pages/Interview/index.js
@@ -4,30 +4,17 @@ import { ReflexContainer, ReflexSplitter, ReflexElement } from 'react-reflex';
 
 import 'react-reflex/styles.css';
 
-import styles from './index.less';
-import ListInput from './ListInput';
-import EditableTagGroup from './TagInput';
+import styles from './styles.less';
+import ListInput from '../../components/ListInput';
+import TagInput from '../../components/TagInput';
+import Header from '../../components/Header';
 
 export default class ReflexBasicSplitterDemo extends React.Component {
   render() {
     const minPanelSize = 150;
     return (
       <div className={styles.container}>
-        <div className={styles.header}>
-          <div className={styles['tool-container']}>
-            <div>icon1</div>
-            <div>icon2</div>
-            <div>icon3</div>
-          </div>
-          <div className={styles['tool-container']}>
-            <div>记录</div>
-            <div>标签</div>
-          </div>
-          <div className={styles['tool-container']}>
-            <div>button1</div>
-            <div>button2</div>
-          </div>
-        </div>
+        <Header />
         <div className={styles.card}>
           <ReflexContainer orientation="horizontal">
             <ReflexElement className={styles['up-container']} minSize={minPanelSize}>
@@ -39,7 +26,7 @@ export default class ReflexBasicSplitterDemo extends React.Component {
               </div>
             </ReflexSplitter>
             <ReflexElement className={styles['down-container']} minSize={minPanelSize}>
-              <EditableTagGroup />
+              <TagInput />
             </ReflexElement>
           </ReflexContainer>
         </div>
