+++ src/pages/Interview/styles.less
@@ -8,22 +8,7 @@
   align-items: center;
 }
 
-.header {
-  background: @base-white;
-  padding: 1.6rem;
-  color: @text-color;
-  border-bottom: 1px solid @white-3;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-}
-
-.tool-container {
-  display: flex;
-}
-
 .card {
-  background: @base-white;
   box-shadow: @shadow-2;
   margin: 2rem;
   height: 100%;
@@ -32,7 +17,8 @@
 }
 
 .up-container {
-  //background: @red-2;
+  background: @base-white;
+
 }
 
 .down-container {
